Issue #4818 Megawidget Scrolling and Size Change Bugs
Made all container megawidgets respond properly to their child
megawidgets resizing themselves.

Made Composite, Group, DetailedComboBox and TabbedComposite all have the
new scrollable property. If true, they provide scrollbars when their
child megawidget panels grow too large to fit within the parents'
bounds. This means that when the metadata in the HID is for example
within a TabbedComposite (e.g. for FL.W hazards), the latter will have
internal scrollbars when necessary, instead of the entire
TabbedComposite scrolling as before (which led to the tabs at the top of
the megawidget scrolling out of site, making the arrangement extremely
annoying to use with anything but a gigantic HID).

Changed the ExpandBar megawidgets to remember their pages'
expanded/collapsed states when the megawidgets are part of the metadata
panel within the HID. Thus, if the HID is closed while a hazard event is
showing that includes ExpandBar megawidgets, and is subsequently
reopened to show that same hazard event, the ExpandBar megawidgets
restore their previously visible pages to be expanded once again.

Laid groundwork for eventually allowing megawidgets to save and restore
their visual state (e.g. position of their scrollbars, etc.). This is
complete as of this code review for the various container megawidgets,
but not others such as CheckList.

Fixed bug in spinner megawidgets that caused them to act as if they
should expand horizontally whenever they were configured to show a scale
widget as part of their layout, regardless of their expandHorizontally
property's value.

Fixed bug in DetailedComboBox causing the label to the left of the
combo box to not always get displayed.

Made minor changes in response to Mike's comments.

Change-Id: Ibad3956e128dd1f650ac01a4a475cda788e865ee