Issue #4124 Selected Time Changes to Encompass All Selected Events
Changed selected time mode to be mutually exclusive: If in single mode,
then a single selected time thumb is shown; if instead in range mode,
two thumbs defining a range are shown. In each mode, the thumb(s) for
the other mode are hidden.

Added code to ensure that whenever the hazard event selection changes,
the selected time is moved or expanded to intersect all selected hazard
events' time ranges, so that all selected hazards show in the spatial
display. Also added similar code for when the time range of a hazard
event changes.

To test, in addition to the usual tests, create some events using the
River Flood Recommender and try selecting different events; as they are
selected, the selected time changes as necessary to ensure it falls
within the selected event's time range. Try also multi-selecting events;
the selected time changes, and/or expands into a range, to continue
ensuring that it intersects all the selected events' time ranges.

Change-Id: I852906cd933471f5d4b2d9cbd3326f7a54075a4a