Issue #4124 Console Sorting
Fixed a number of bugs that caused sorting to work inconsistently at
best within the Console. For example, sorting by ETNs when some events
had ETNs and some did not resulted in the blank ETNs being above the
ones that had non-null ETNs, regardless of which direction the sort was
in. This required a lot more changes than one might suspect; part of the
problem was duplicate notifications being received because we're still
working with a mixture of two different notification propagation
systems. It also required that the Settings/ObservedSettings classes be
reworked so that both implement ISettings, and ObservedSettings
decorates Settings, similarly to the way IHazardEvent and
ObservedHazardEvent works.

Added the ability to sort by more than one column, with columns
prioritized (i.e. sort by column A, and then for any that have equal
values in A, sort by their values in B, etc.). The menu structure
supports only two-column sorting (primary and secondary), but this is as
per the requirements; the rest of the sort code has been written to be
able to sort using N different columns. Thus, if we ever need tertiary
sorting, for example, it should be trivial to add (minor GUI changes
only).

Added the capability of sorting by hidden columns. Previously, only
visible columns could be used for sorting. Hidden columns may be
specified for sorting (primary or secondary) via the column header
menus.

See the "Sorting Tests" section of the following document for how to
test this changeset:

  https://docs.google.com/a/noaa.gov/document/d/1wv8I233xAdmDZXDP4b6f63_LhH8xEZs9kt6uBsLHXrU/edit

Manually rebased to fix merge conflicts.

Change-Id: I48b50988dfc9ffdd1e1e7765450276678222ad1a