Issue #3783 HZ behavior for selecting the end time of a river FL.W
Sets Until Further Notice to true for Time Range End and Fall Below Time if
river gauge never falls below flood stage in forecast hydrograph.  Also sets
end/fallBelow times to last value in forecast hydrograph in in this situation
if the user unchecks the UFN box.

Additional fix to RiverFloodRecommender.py moves RiverForecastPoints instantiation
to the execute() method so changes to the hydro database are reflected in subsequent
runs of the recommender

Change-Id: If50126c87ab74fe5594f09d59f6bd0901fa3e8b3