Issue #4959.  Added some of the capability to right-click to add/remove UGCs.  Completion in a follow-on code review.
One pathway not yet working is the if the forecaster
uses selectByArea for counties to create the hazard geometry initially and t
uses MB3 to add/remove UGCs.  There is a mismatch between the
geometry create via selectByArea and the geometry created when
the user adds/removes.  This leads to messed up hazard geometries.

Also made a downpayment on RM 5952 (incorrect UGC and hatching designation).
Made some corrections to configuration in HazardTypes and modified
the code accordingly.