Issue #2331 Start/End Time Editability Restrictions (part 1 of 2 changesets)
Added ability for SWT custom widgets (multi-value scale and ruler) to
have individual min-max boundaries set for their different thumbs, so
that for example one or more thumbs can be limited to only go up or down
to a certain value, with each thumb having different boundaries.

Added individual minimum and maximum allowable boundaries for the first
state identifier's value for TimeRange megawidgets, and for any state
identifier's value for TimeScale megawidgets, so that such a megawidget
could have it start time limited to go no earlier than a certain
timestamp, for example.

Fixed a small bug in the manual megawidget tests that caused
icon-equipped buttons in megawidgets to display black rectangles instead
of their icons.

Changed the megawidget test application to ensure that any
timestamp-manipulation megawidgets with scale widget components have
their values' thumbs visible along the length of their scales.

Note that this changeset does not actually add the business logic to
Hazard Services to limit the start or end time editability; it is
instead designed to provide the underlying functionality needed by the
megawidgets and SWT custom widgets to support the use of boundaries on
individual values.

For testing, see the Notes section of the Redmine issue, not the Test
Steps section, as those are designed to test the limitations themselves,
and will apply in the second changeset, not this one.

Fixed egregious and boneheaded copy-and-paste-caused bug from the first
couple of patch sets that caused the mixed hazard story to fail. Mea
culpa.

Change-Id: I0f9798f95235a3ca37521592ff207ea759d64192