Issue #3626 Provide Hazard Type to Recommenders
When a recommender is run as the result of a hazard-type-first dialog
invocation, the hazard type itself is now passed to the recommender, so
that it knows what sort of event(s) it should be creating. The recommenders
have also been modified to accept this information and use it as appropriate.

Change-Id: Ie55ce4f3942d8e597cbaab3ee82c0c5398a43543