Issue #5952.  Further progress on UGC/hatching issues (still another code review to come later).
Fixed a bug in the hatching for polygon-based hazard types.  Due to
round-off error, neighboring counties were being hatched improperly.
Mitigation approach is to set a minimum inclusionPercentage (something
like .0001).

Added area based inclusion thresholding.

Added exception handling for when product generation fails due to a bug
so that the preview/issue buttons will be not be left disabled.

Change-Id: I511893883d0f03943f866ec694e13594e2078de8