Issue #6370 Right-Click Column Header Menu Not Showing All Columns
Fixed bug that manifested itself when, for example, the user started
H.S. from the D2D perspective (thus using the TOR settings), then
switched to the Flood settings and right-clicked on the Console column
headers. The resulting menu would not include the Stream or Point ID
columns, even though they are available in the Flood settings.