Issue #6393 Configurable Title Text in HID Tabs
Added option to StartUpConfig.py that allows the specification of which
properties of a hazard event should be shown as the title text for that
event in the HID tab headers. Title text may be made up of one or more
properties of a hazard event, such as event ID, type, and any attributes
the event has.

Change-Id: I0c88e3e446ced3037eba570b116df1f34e7a25ce