Issue #4756 Persisting Display Properties Between Metadata Refreshes
Changed megawidgets so that when they are refreshed, they keep their
display properties across the refresh when possible. For example, a Text
megawidget checks to see if it has the same state (text contents) as it
did before the refresh; if it does, it then restores the scrollbar to
the same position it had before, selects any text that was selected
before, etc.