Issue #3850 Limiting Hazard Type Choices
Changed the Hazard Type pulldown in the HID so that it only allows the
selection of hazard types that make sense for the event being displayed. 

For pre-issued events, this means that any event type marked in
HazardTypes.py as requiring a point identifier (currently, these are
FL.A, FL.W, FL.Y, and HY.S) cannot be selected unless the event contains
a point identifier.

For issued events, the only event types allowed (other than the
currently selected type) are those that are listed as potential
replacements for the current type in HazardTypes.py.

Note that this code review also addresses issue #4210 (smart choices for
hazard types) since the implementation is largely shared.