Issue #7102 Start/End Time Limitations
Fixed at least one of the two bugs causing exceptions in the start/end
time limitation code introduced as part of Redmine issue #2331. Both of
the bugs should be fixed, but since Chris Golden was unable to reproduce
the exceptions seen in the field, he cannot be sure. Fallback code has
been included in case the one bug still manifests itself, so that such
manifestations will not leave Hazard Services in a bad state.

Also fixed bug that caused events that went from ISSUED to ENDED status
directly (no intermediate ENDING status) to still allow their start and
end times to be manipulated. When ENDED or ENDING, this should not be
allowed.

Fixed bug that caused an exception to occur when the table does not
contain a row for a hazard event that has experienced a time range
limitation change.

Also fixed design flaw with start/end time limitations. The original
limitation code advanced the start time of a hazard event to the current
time each time it was reissued (if it was less than the current time at
issuance). This behavior was incorrect; it should not be advanced after
the first issuance. After issuance, the start time is also not limited
at all if the hazard type does not require the start time to be the
current time.

Change-Id: I23048c154eeff0767edcb5dace16635e0bf77fec