Issue #6873 Mouse Wheel Fixes
Added code to stateful megawidgets that were consuming mouse wheel
events to cause them to instead pass said events up to the nearest
scrolled composite ancestor widget they can find. This will allow a
panel of megawidgets to be scrolled using the mouse wheel without the
scrolling suddenly switching to a megawidget when the mouse cursor ends
up (as a result of the panel scrolling) over a megawidget's UI elements.

Change-Id: Ia5f28acb54bb6b47941921b1ca991d4f0137dfa1