Issue #4162 Read-Only Table Megawidget
Added read-only table megawidget to the megawidgets toolkit, allowing
tabular data to be displayed within megawidgets. The megawidget is used
in place of labels (which were not well-suited to this task, and were
being used only as a placeholder until we had a table megawidget
available) for such data display in the FL.W metadata. Also fixed bug
noticed in code review.

Change-Id: I68ce498f0658506e8583d030d2a89717947b1e36