Issue #4162 Read-Only Table Megawidget
Added read-only table megawidget to the megawidgets toolkit, allowing
tabular data to be displayed within megawidgets. The megawidget is used
in place of labels (which were not well-suited to this task, and were
being used only as a placeholder until we had a table megawidget
available) for such data display in the FL.* metadata.

Note: This changeset has already been reviewed, merged, and then
reverted because it was making merging 15.1.1 changes into 16.1.1
difficult. This version makes the metadata changes in CommonMetaData
instead of MetaData_FL_W, since much of the functionality in the latter
file was moved to the former as a result of the 15.1.1 to 16.1.1 merge.