Issue #7271 Render Rise/Crest/Fall Values Read-Only in HID
Changed the HID to use read-only text fields to display the
rise/crest/fall values. With the introduction of the graphical
rise/crest/fall editor, these values may now be set to special "missing"
value, which does not translate nicely into the a time scale megawidget
and additionally was leading to exceptions being thrown; thus the
urgency of the change.

In the process, changed the River Flood Recommender to provide values
for rise/crest/fall when creating FL.Y-type events, which previously
always had values of 0 for each of these attributes.

Note that further work will be required to complete the rise/crest/fall
implementation, but such work falls outside of the scope of this task,
since this code review is primarily intended to prevent exceptions from
occurring related to rise/crest/fall. These are noted in Redmine issue
#7560.