Issue #6935 HID Enhancements
Added "prompt text" parameter to Text megawidgets, allowing them to show
gray text indicating what is to be typed into them; this text disappears
as soon as the user puts one or more character within the field. This is
to resolve FAT issue 7.14.

Added ability to specify which ExpandBar megawidget's pages are to be
expanded (visible) when it is created, and for symmetry's sake, which
TabbedComposite megawidget's page is to be topmost (visible) when it is
created. This change is to resolve FAT issue 7.34.

Fixed bug found while developing the above; the bug caused duration list
for hazard events (displayed in HID) to lag behind the currently
displayed event.

Note that FAT issue 7.76 is not addressed by this changeset.

Change-Id: I5cc8a7e19706e2591ef697d97dbeb6bc96032ea1