Issue #6979 Finish work on product correction:
1. Created a update method in the generator that updates previously
generated dictionaries instead of recreating them. This is needed
because staging dialog info and correction info was being lost when the
previously generated dictionary is not used.
2. Added save button back to Product Editor to allow users the
choice to save or not.
3. Changed formatters to take a dictionary of editable parts that
come from the Product Editor when it regenerates. This avoids having
to save every edit to the productText table.
4. Made LabelMegawidgets stateful so that they could be updated when
doing corrections.
5. Added column to productData table to store the editableEntries upon issue.
This is needed along with the productText table. Since edits could be made
after the product has been issued. Then users could attempt to bring the product up
for a correction.

Change-Id: I78556b6e4f128a35c6b41a600c4a1f70706ca99f