Issue #7376 Changes for optional product parts and removed most framed text:
1. Hooked up the "required" functionality of the Product Editor. Meaning
certain KeyInfo objects must contain values before the issueAll button
is enabled. These are marked with a "*" in the labels.

2. Removed all framed text for optional product parts. Since it is optional
the text is ommited from the product. Users can still add something to the
product editor for the part and then it will get added to the product.

3. Misc bug fixes.

Note: The only frame text that remains is for required product parts or if
metaData fields fail to get replaced by HID values.