Issue #4245 Time Ruler in Hazard Detail View
Added a "time ruler", similar to the one found above the time scale
widgets in the Console, to the Hazard Detail View's start/end time
widgets. This gives the user context for what values the thumbs in the
time slider represent, and also allows the time to be panned or zoomed
from within the Hazard Detail View, not just the Console. As before, if
the StartUpConfig option "showHazardDetailStartEndTimeScale" is set to
False, no scale bar or time ruler is shown.

Change-Id: I24a49e6e92f73321dbb9957c467b12294473b18e