Issue #4123 Changes to Number-Choosing Megawidgets
Added two new megawidgets, a fraction-choosing range and an
integer-choosing range. Each allows the selection by the user of a range
of numbers, via individual choosing of lower and upper bounds. Like the
spinner megawidgets, they may be configured to show a scale bar beneath
them, allowing the user to drag the thumbs left and right to select the
range boundaries.

Also changed the spinner megawidgets to use (when showing a scale) the
same custom SWT widget (multi-thumbed scale) as the range, so as to
allow more sharing of code and to maintain a consistent look and feel
between the spinner and range megawidgets.

Change-Id: I854af59f15ac516c8a6a5db0ae8507a196562153