Issue #9617 Fixes for Code Merged Previously
Fixed problems introduced by code for this ticket previously that messed
with the megawidget framework. Also renamed the new combo box option
from "readOnly" to "allowNewChoice", as this better describes the
option.

Removed unneeded and unwarranted dependencies from the megawidget
project's manifest.

Removed product editor specific code from the megawidget framework,
putting copies of the parameters editor factory and one accompanying
interface in the product editor project to allow the undoing of the
messy changes that had been made to the megawidget framework, but still
allowing the product editor to retain the benefits of said changes.

Fixed megawidget test code.