VLab Issue #12762 Run Recommenders In Response to Triggers
Moved all recommender management code into new recommender manager class
in session manager, consolidating it in the proper place.

Added ability to trigger the running of arbitrary recommenders in
response to changes to hazard events (both first-class fields such as
time range, and also arbitrary metadata attributes).

Also added ability to run arbitrary tools (recommenders or product
generators) at regular intervals in the background.

Changed recommender result handling so that if the recommender execution
does not result in any hazard events, the user may or may not be warned
depending upon whether the recommender is meant to run in the foreground
or not, since having recommenders running behind the scenes at regular
intervals should not result in the user being notified of no new events.

Fixed bug in hazard metadata accessor that did not provide metadata
megawidgets for the HID in cases where the hazard type has no
significance.

Fixed compilation bug in flash flood recommender causing a Python JEP
error to occur. Also changed recommender execution code to use only
one thread, not two, as a race condition (long-standing but until now
not detected) was intermittently causing two instances of a recommender
to be run for different facets of the recommendation process (getting
dialog info, executing the recommender itself, etc.), which meant the
wrong instance might be called with information it needed not available
to it because said information was part of the other instance.

Change-Id: Ibe18de1e7e20f910eb38ec700e3475da9d433004