VLab Issue #16145 Fix GeometryCollection serialization bug
Fixed long-standing GeometryCollection serialization bug. It turned out
that serialization of GeometryCollection objects within the "geometry"
first-class field of a hazard event were not a problem; rather, the
problem occurred because hazard attributes sometimes have serialized
Geometry objects as values. This fix should allow for both Geometry
values in attributes, or lists, sets, or maps in which the values are
themselves Geometry objects.

Note that bringing in this changeset requires a rebuild of edex,
followed by a restart of both the ebxml-registry and edex request.

Change-Id: I77fd33c1a5e465113a055aa10f64e14bbdbbba48