VLab Issue #15676
Step one of visual features additions. This allows visual features to be
displayed, though not perfectly, and certainly not in a manner that
allows them to be editable; that will come in a follow-on code review.

Also added ability for recommenders to specify that they only want in
their input EventSet the hazard event whose modification triggered their
execution (if appropriate), and/or to indicate they only want the input
EventSet to include hazards with certain types.

Added additional execution context information for recommenders being
run, including the origin of the change (if a hazard event change
triggered the execution), the current selected time, etc.

Change-Id: I4c408a78530115dec97e9b64f82573bf82ae2f1b