VLab Issue #15676 Visual feature editing and moving
Added the ability to move entire visual features, and to edit visual
features' geometries when the latter are multi-point (e.g. polygons).
This mirrors the capabilities the spatial display provided already with
hazard event geometries.
    
Fixed a number of bugs in the spatial display along the way.
    
Wired together visual feature changes to trigger recommenders as
appropriate.