VLab Issue #15931 Graph Megawidget (part 2, bug fix)
Fixed bug in the Graph megawidget (or actually, the underlying graph
custom SWT widget) that threw a null pointer exception when the plotted
points were set to a new list that was the same size as the old list of
points.

Added antialiasing when painting the widget to reduce jagged edges of
the plotted points and the line connecting them.