VLab Issue #15931 Graph Megawidget (part 2, second bug fix)
Fixed typo in CommonMetaData specification of Graph megawidget that
caused the intended "sendEveryChange" option to be ignored.

Fixed bug in Graph megawidget that caused starting state to be a
reference to an external object, instead of a copy of the latter,
meaning that an ObservedHazardEvent was sharing an ArrayList with the
Graph megawidget, which in turn meant that attempts to set the hazard
attribute of the event failed (and did not generate any "attributes
modified" notifications) because the event already had the updated
value.