VLab Issue #17467 Removed inappropriate ignoring of added events
Changed session event manager to no longer refuse to add or modify
events when the added event (or the new version of the modified event)
would be filtered out by current settings filters.

To test, maker sure you rebuild edex. Then start Hazard Services and
give it half a minute for the Convective Recommender, etc. to finish
running. Then reset events and close Hazard Services. (This simply
serves to ensure the slate is clean when actually testing.)

Now without restarting CAVE, start H.S. again, and let it run. It will
start at the canned time of 11-11-2015 18:36Z, and the clock will run
forward (i.e. CAVE clock is not frozen). Make a note of the event
identifiers (not the Object IDs!) of the hazard events that the
Convective Recommender initially generates. No need to select any
events, etc.; just let CAVE run.

After maybe six or seven minutes, as the CAVE clock marches forward, one
of the events will be changed to "ended" status, and a new event will
replace it. In the Console you may miss this as the change is subtle;
since the ended event will be filtered out (invisible), it'll go from
three events to three events, but one of the events will now have a
different ID. (This is the event ID needs to be noted in the beginning.)

If this happens, the test passes. (Previously, if the "ended" status was
filtered out by the current settings, as it is in this changeset, the
attempt to change the event to "ended" status would be ignored by H.S.;
the event would remain "potential" or "pending" or whatever it was
before.)