VLab Issue #15676 Fixed bugs with visual features and spatial display
Added code to deserialization process for visual features to (a) check
for bad geometries and throw an error if one is encountered, and (b)
warn via stderr when a visual feature's property's time range has
boundaries that are not rounded to the minute.

Added a toString() method to visual features to indicate the time
range(s) in which they have valid geometries, for debugging purposes.

Changed Python hazard event class to return an empty list instead of
None when getBaseVisualFeatures() or getSelectedVisualFeatures() is
called and there's nothing to return.

Made the spatial display more generous in determining whether a user's
click and drag landed on a shape, and also changed the algorithm used to
determine the vertex being edited (if any) to use the closest vertex to
the point on the display clicked by the user, not the first vertex
within a certain distance of that point.