VLab Issue #16885 User-Owned Column in Console Table
Added ability to show a column in the Console table for the "user-owned"
hazard attribute. This column is shown by default within the Prob_WFO
setting.

To test, start H.S. and run the Convective Recommender. The events
created should all be marked with a "no" in the "User-Owned" column of
the Console's table. Now select one of them, and change its "User-owned"
checkbox in the HID to checked; the corresponding cell should update to
"yes" in the Console.

Now create a new hazard event with no type. It should have "N/A" as its
entry in the "User-Owned" colum of the Console's table.

Now try sorting by the new column. Sorting one direction should put
"yes" at the top, then "no", then "N/A". Reversing the sorting should of
course make the ordering "N/A", "no", and "yes".

And obviously it should run through the Bare Bones Story without a
problem.