VLab Issue #15931 Graph Megawidget (part 3)
Added ability to allow the user to draw points directly on the Graph
megawidget (and the underlying graph SWT custom widget) if the Graph is
cleared of points, and then the user clicks, drags, and releases to draw
a line along which points should lie. Prompt text appears in the body of
the graph if it is cleared and this option is enabled, so that the user
knows what can be done.

Added code to ensure that focus loss while a mouse operation is ongoing
does not leave the Graph in an undefined state.

Added use of mouse cursors to indicate capabilities when the user moves
the mouse over the Graph.

Added code to the PHI-related interdependency script of CommonMetaData
to cause the Graph to be cleared of points when the user invokes the
"Draw" button; this will put the Graph in a state where it can accept
user drawing of points. Also disabled the "Draw" button, however, since
the Swath Recommender has not yet been changed to accommodate user
drawing of points.

For testing, there is little to do since the new drawing functionality
is not available within H.S. due to the disabled "Draw" button. However,
testing the Graph as before is a good idea, and note that the cursors
change as the user hovers them over the existing points to indicate that
they are capable of being dragged, except the rightmost point which is
uneditable.