VLab Issue #15931 Graph Megawidget Drawing and Pickling Issued Event
Added code to the Swath Recommender to handle the Graph megawidget going
into user-draw-points-on-graph mode, and then coming out of it with new
points (or that mode being canceled by other changes made to the hazard
event).

Changed metadata generation code to enable the Draw button for the
Graph.

Added code to "pickle" a convective hazard event upon issuance (from
Kevin M.).