VLab Issue #18094 Expanded valid values for hazard attributes
Changed hazard attributes to allow them to be successfully serialized
and deserialized when their values are collections or maps holding
heterogeneous types, and/or nested collections or maps. Previously such
values could not successfully be deserialized, leading to lists of nulls
and other such nonsense when a hazard event was read back from the
database, and thus causing havoc.

Fixed a bug that caused the auto-running of time-interval-driven
recommenders to look like it was enabled upon H.S. startup, but not
actually occur. Now, when H.S. starts up, the auto-running proceeds
correctly.

Fixed a bug in the event manager that caused exceptions when issuing
events without VTEC codes (e.g. PHI event types) because such codes were
expected by the manager.

Change-Id: I90358a74cf6c3f1b78bf1652b1b8186acae3730c