VLab Issue #18267 Option for start time to be earlier than current time
Added option for a hazard type to have its start time be unrestricted,
that is, it can be earlier than the current time. (Previously that was
not allowed except for issued events, and for them the start time could
be no earlier than the issue time.)

Fixed bugs in the recently expanded hazard attributes:

Sets and geometries that were the value of the attribute (not nested
within the value) to not be deserialized correctly.

Made collections and maps with values that include null be valid, but
not "slottable" (previously they were considered invalid). The latter
means that they are not given slots for their individual values by the
slot converter when being placed in the ebXML registry, but are instead
simply serialized as JSON.