VLab Issue #18376 Product generators/recommenders throw numpy errors
Changed the Jep usage for both product generators and recommenders so
that the Jep instance for each is a singleton that is created the first
time H.S. starts up in a CAVE session, and then not shut down until CAVE
exits. This way, it is reused for each successive H.S. session within
the same CAVE session.

This is being done to avoid numpy errors when H.S. is restarted in the
same CAVE session, since numpy does not play nicely with Jep if the
latter is closed and then a new Jep instance with numpy usage is
created.