VLab Issue #15676 Visual Feature Thrift Serialization/Deserialization
Changed visual feature class to properly support Thrift serialization
and deserialiation, which in turn means no more strange
visual-feature-related errors in the ebXML registry, and thus the
sharing of events between different Hazard Services instances connected
to a common edex. Now, when one stores an event (issuing, etc.), the
others update to reflect the new version of the hazard event.

Change-Id: I6bbfc88dcc9b46fdd51df8267c71215efc4ae17c