VLab Issue #18515 Optionally Deselect Hazard Events After Issuance
Added code that allows settings to contain a new flag,
"deselectAfterIssuing", which if set to true causes Hazard Services to
deselect an event following its (re)issuance.

Eventually this should be replaced by the ability of the product
generators to indicate whether they want a hazard event that was just
issued to be selected or not, but for now, this will provide what the
HWT needs.

Also fixed a bug in the HazardEvent Java class (used by H.S. when
storing hazard events to the database, etc.), that allowed multiple
attributes with the same key to be saved as part of the set of
attributes, which was quite problematic. It also caused attributes that
the software thought had been removed to hang around.