VLab Issue #15676 Fix for storeEvents() error with larger PHI events
Fixed the storeEvents() bug, at least for now, but removing visual
features from hazard events prior to storing them. It is expected that
visual features are generated by recommenders when hazard events that
require them are loaded from the database (either when starting H.S. or
when a newly (re)issued, ended, etc. hazard event arrives).

Added new "database" originator in Java code, so as to differentiate
between database-sourced changes in notifications versus
side-effect-caused changes. Also added "database" as an origin for
recommender triggers, so that the Swath Recommender can know that if it
is triggered by a database-originated change, it should always build the
visual features for the changed event.