VLab Issue #19464 Add visual cues to required fields in product editor
Added the ability to specify the color of a Label megawidget, and then
changed the product editor to use composites of two labels, one black
and one red, for any required fields. This is being done to allow the
latter to visually stand out and remind the user that they require
attention.