VLab Issue #19537 Move selection-based visibility into visual features
As part of the spatial display refactor, visual features are being
changed so that each one knows the constraints placed upon its
visibility with respect to the selection state of the object it
represents.

Uncommented the code used to generate history polygons in the Swath
Recommender (these were commented out temporarily for the HWT).

Change-Id: Id562d05b7337c026823959f293cf4640d61af244