VLab Issue #19537 First step of Spatial Display refactor
Partly refactored the Spatial Display. This changeset focused primarily
upon removing all special-case code from the Java core that handled the
storm track tool functionality, replacing it with use of the generic
visual features now available to all recommenders, and adapting the
storm track recommender to the new interface.

As part of the above, the ability to draw a star symbol instead of a
point for a visual feature was added, as well as the ability to specify
a visual feature as being drawn topmost in the z-order. Furthermore,
visual features may now be built so as to have label text identical to
the hazard event they represent. Finally, label positioning (offset and
direction), which was previously configurable but had no effect on the
rendering of the visual feature, now affects the positioning of the
label text as it should.

Added the ability to allow a recommender to provide pre-execution visual
features that the Spatial Display uses to gather spatial information
from the user. (This replaces the old spatial info code, which again was
hard-coded to work only with the storm track tools.)

Adapted the storm track tool to the new generic visual features
interface, merging the modify storm track tool into it in the process.

Previously, hazard events with visual features never displayed hatching;
this has been changed, so that there are now three choices for hatching
(configurable by hazard type): none, intersecting-regions, and
WarnGen-style. Visual-feature-equipped hazard events may use any of
these.

Replaced uses of Python's datetime.fromtimestamp() with
utcfromtimestamp() in many places, since that is more appropriate.
(There are still instances of fromtimestamp() being used where there was
uncertainty as to which was appropriate.)

Change-Id: I6bc85b1060ff46994bf5653a3d504e54413944e3