VLab Issue #19464 Changed storage of colors from maps to Color objects
In response to code review comments in the H.S. Operational side,
changed the label specifier and megawidget to use RTS Color objects for
storing colors instead of maps of floating-point values. The latter are
still expected from the JSON; they are merely converted to Color objects
when stored internally within the megawidgets.

Deprecated the methods that get colors as maps, as their usage should be
replaced in the Graph megawidget in the future; but since Graph isn't in
Experimental, it wasn't worth doing this work right now under such
severe time constraints.

Change-Id: I9485c42bb1512b51f24d80e41a6c0ff043ec1214