VLab Issue #19537 Second step of Spatial Display refactor
Extensively refactored the Spatial Display components to follow the H.S.
MVP architecture. The work is not complete, as there is still too tight
a coupling between the presenter and the view, but most of the rest is
done.

In the process, fixed several bugs, most notably the undefined state
behavior that H.S. would exhibit when the user was creating a new hazard
event via geometry creation or select-by-area, then selected another
spatial input mode in the Console toolbar before completing the
creation.

All displayable objects that are passed from the presenter to the view,
and from there to the display, are now SpatialEntity instances, whether
for hatching, for representing the base geometry of a hazard event,
visual features of a hazard event, or visual features used by tools for
spatial information gathering. In the process, a lot of code was removed
that handled hazard events as displayables directly in the display and
the mouse handlers, which was problematic from a separation-of-concerns
perspective and was responsible for a lot of the spaghetti-like nature
of the code.

Visual features now support two different fill styles (solid and
hatched).