VLab Issue #20755 Recommenders ask that their events be persisted
Added ability for a recommender to request that the events it createsd
or modifies (and returns in its resulting event set) be saved to the
database. A recommender may do this by adding the attribute
"saveToDatabase" to the resulting event set, with a value of True.

Fixed a bug in save-events code that neglected to strip visual features
and other elements of a hazard event that are database-unfriendly before
saving.