VLab Issue #19924 Trigger tools from Time Match Basis changes
Changed the data-layer-update tool triggering code to instead only
trigger when the Time Match Basis (TMB) product times changed (which
will only occur in D2D). Also changed the recommender code so that the
entire list of TMB data times is provided to recommenders that ask for
it, not just the latest time.

Change-Id: Ic7f9d12180943ef5dec953d37b4af4b9f16bd620