VLab Issue #18376 Numpy errors during PHI metadata generation
Fixed bug caused by bad interactions between the Python numpy library
and Jep, which manifested itself as exceptions being thrown and metadata
generation not completing successfully when a PHI hazard event was
created or viewed in one session of H.S., and then in another session
within the same CAVE session, a PHI hazard event was created or viewed
again. The solution, as with product generation and recommenders, is to
make the the metadata generation Jep instance a singleton that gets
reused from session to session as long as CAVE is still running.

Fixed exceptions thrown having to do with setTimeRange() when, in a
second H.S. session, a hazard event was created and saved. This was
caused by zombie H.S. objects sticking around after H.S. shutdown, and
reacting to database changes when in an undefined (shut down) state. The
solution for now has been to keep a flag indicating whether the session
event manager has been shut down; the long-term solution is to determine
what is causing the session manager components, many UI components, etc.
to not be garbage-collected. This will be followed up on against Redmine
issue #21271.

Removed some unneeded code and did some work on removing extraneous
references to H.S. objects to move toward the goal of #21271.

Change-Id: Ice47b8debe3a89e6d7173c3df1fc47b2b3e30467